requirements.txt
crewai>=0.30.11
crewai-tools>=0.1.7
hyperbrowser
langchain-community
streamlit
python-dotenv